tailwind fixed, creating components

+++ example-app2/app/Http/Controllers/Tweet/DeleteController.php
@@ -4,13 +4,18 @@ namespace App\Http\Controllers\Tweet;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use App\Services\TweetService;
 use App\Models\Tweet;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class DeleteController extends Controller
 {
-    public function __invoke(Request $request)
+    public function __invoke(Request $request, TweetService $tweetService)
     {
         $tweetId = (int) $request->route('tweetId');
+        if(!$tweetService->checkOwnTweet($request->user()->id, $tweetId)) {
+            throw new AccessDeniedHttpException();
+        }
         $tweet = Tweet::where('id', $tweetId)->firstOrFail();
         $tweet->delete();
         return redirect()
